+++ src/main/java/chess/Application.java
@@ -3,8 +3,8 @@ package chess;
 import chess.controller.ChessController;
 
 public class Application {
-    public static void main(String[] args) {
-        ChessController chessController = new ChessController();
-        chessController.run();
-    }
+	public static void main(String[] args) {
+		ChessController chessController = new ChessController();
+		chessController.run();
+	}
 }
