+++ src/main/java/chess/domain/ChessGame.java
@@ -1,42 +1,43 @@
 package chess.domain;
 
-import chess.domain.piece.Color;
-import chess.domain.piece.kind.Piece;
+import static chess.domain.piece.Color.*;
 
 import java.util.Map;
 
-import static chess.domain.piece.Color.BLACK;
-import static chess.domain.piece.Color.WHITE;
+import chess.domain.board.Board;
+import chess.domain.board.Point;
+import chess.domain.piece.Color;
+import chess.domain.piece.kind.Piece;
 
 public class ChessGame {
-    private final Board board;
-    private Color currentColor = WHITE;
+	private final Board board;
+	private Color currentColor = WHITE;
 
-    public ChessGame() {
-        this.board = new Board();
-    }
+	public ChessGame() {
+		this.board = new Board();
+	}
 
-    public void playTurn(Point source, Point target) {
-        board.movePiece(source, target, currentColor);
-        this.currentColor = switchTurn();
-    }
+	public void playTurn(Point source, Point target) {
+		board.movePiece(source, target, currentColor);
+		this.currentColor = switchTurn();
+	}
 
-    private Color switchTurn() {
-        if (this.currentColor.equals(WHITE)) {
-            return BLACK;
-        }
-        return WHITE;
-    }
+	private Color switchTurn() {
+		if (this.currentColor.equals(WHITE)) {
+			return BLACK;
+		}
+		return WHITE;
+	}
 
-    public boolean isNotEnd() {
-        return board.hasBothKings();
-    }
+	public boolean isNotEnd() {
+		return board.hasBothKings();
+	}
 
-    public Score calculateScore() {
-        return board.makeScore(currentColor);
-    }
+	public Score calculateScore() {
+		return board.makeScore(currentColor);
+	}
 
-    public Map<Point, Piece> getBoard() {
-        return board.getBoard();
-    }
+	public Map<Point, Piece> getBoard() {
+		return board.getBoard();
+	}
 }
