+++ src/main/java/chess/domain/piece/Color.java
@@ -1,14 +1,14 @@
 package chess.domain.piece;
 
 public enum Color {
-    WHITE("WHITE"),
-    BLACK("BLACK"),
-    NOTHING("NOTHING");
+	WHITE("WHITE"),
+	BLACK("BLACK"),
+	NOTHING("NOTHING");
 
-    Color(String colorName) {
-    }
+	Color(String colorName) {
+	}
 
-    public boolean isSameAs(Color color) {
-        return this.equals(color);
-    }
+	public boolean isSameAs(Color color) {
+		return this.equals(color);
+	}
 }
