+++ src/main/java/chess/view/OutputView.java
@@ -1,52 +1,52 @@
 package chess.view;
 
+import java.util.Map;
+
 import chess.domain.ChessGame;
-import chess.domain.Point;
+import chess.domain.board.Point;
 import chess.domain.Score;
 import chess.domain.piece.kind.Piece;
 
-import java.util.Map;
-
 public class OutputView {
-    public static final String START = "start";
-    public static final String END = "end";
-    private static final String START_GUIDE_MESSAGE = "체스 게임을 시작합니다.";
-    private static final String START_INPUT_GUIDE_MESSAGE = "게임 시작은 " + START + ", 종료는 " + END + " 명령을 입력하세요.";
+	public static final String START = "start";
+	public static final String END = "end";
+	private static final String START_GUIDE_MESSAGE = "체스 게임을 시작합니다.";
+	private static final String START_INPUT_GUIDE_MESSAGE = "게임 시작은 " + START + ", 종료는 " + END + " 명령을 입력하세요.";
 
-    private OutputView() {
-    }
+	private OutputView() {
+	}
 
-    public static void printStartGuideMessage() {
-        System.out.println(START_GUIDE_MESSAGE);
-        System.out.println(START_INPUT_GUIDE_MESSAGE);
-    }
+	public static void printStartGuideMessage() {
+		System.out.println(START_GUIDE_MESSAGE);
+		System.out.println(START_INPUT_GUIDE_MESSAGE);
+	}
 
-    public static void printBoard(ChessGame chessGame) {
-        Map<Point, Piece> pieces = chessGame.getBoard();
-        for (int i = 0; i < 8; i++) {
-            StringBuilder oneLine = new StringBuilder();
-            makeOneLine(pieces, i, oneLine);
-            System.out.println(oneLine.toString());
-        }
-        System.out.println();
-    }
+	public static void printBoard(ChessGame chessGame) {
+		Map<Point, Piece> pieces = chessGame.getBoard();
+		for (int i = 0; i < 8; i++) {
+			StringBuilder oneLine = new StringBuilder();
+			makeOneLine(pieces, i, oneLine);
+			System.out.println(oneLine.toString());
+		}
+		System.out.println();
+	}
 
-    private static void makeOneLine(Map<Point, Piece> pieces, int i, StringBuilder oneLine) {
-        for (int j = 0; j < 8; j++) {
-            oneLine.append(makeOneCharacter(pieces, i, j));
-        }
-    }
+	private static void makeOneLine(Map<Point, Piece> pieces, int i, StringBuilder oneLine) {
+		for (int j = 0; j < 8; j++) {
+			oneLine.append(makeOneCharacter(pieces, i, j));
+		}
+	}
 
-    private static String makeOneCharacter(Map<Point, Piece> pieces, int i, int j) {
-        Point point = Point.valueOf(i, j);
-        return pieces.get(point).getName();
-    }
+	private static String makeOneCharacter(Map<Point, Piece> pieces, int i, int j) {
+		Point point = Point.valueOf(i, j);
+		return pieces.get(point).getName();
+	}
 
-    public static void printScore(Score score) {
-        System.out.println(score.getScore());
-    }
+	public static void printScore(Score score) {
+		System.out.println(score.getScore());
+	}
 
-    public static void noticeGameFinished() {
-        System.out.println("게임이 끝났습니다. 새로운 게임을 시작하시려면 start, 종료를 원하시면 end를 입력하세요.");
-    }
+	public static void noticeGameFinished() {
+		System.out.println("게임이 끝났습니다. 새로운 게임을 시작하시려면 start, 종료를 원하시면 end를 입력하세요.");
+	}
 }
